+++ packages/core-di/src/defaults.ts
@@ -1,9 +1,10 @@
-import { CaptivePolicy, LogLevel, ResolveMultipleMode, RuntimeCaptivePolicy } from '@shellicar/core-di-engine';
+import { CaptivePolicy, Lifetime, LogLevel, ResolveMultipleMode, RuntimeCaptivePolicy } from '@shellicar/core-di-engine';
 import type { ServiceCollectionOptions } from './types';
 
 export const DefaultServiceCollectionOptions: ServiceCollectionOptions = {
   registrationMode: ResolveMultipleMode.Error,
   logLevel: LogLevel.Warn,
   captivePolicy: CaptivePolicy.Disposal,
-  runtimeCaptivePolicy: RuntimeCaptivePolicy.None,
+  runtimeCaptivePolicy: RuntimeCaptivePolicy.Throw,
+  defaultLifetime: Lifetime.Resolve,
 };